resolveCodeAction cleanup
