+++ rosetta-ide/src/main/java/com/regnosys/rosetta/ide/quickfix/RosettaCodeActionProvider.java
@@ -1,16 +1,16 @@
 package com.regnosys.rosetta.ide.quickfix;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.inject.Inject;
 
 import org.eclipse.emf.common.util.EList;
-import org.eclipse.emf.ecore.EObject;
-import org.eclipse.lsp4j.Diagnostic;
+import org.eclipse.lsp4j.CodeAction;
+import org.eclipse.lsp4j.CodeActionKind;
 import org.eclipse.lsp4j.Range;
 import org.eclipse.lsp4j.TextEdit;
-import org.eclipse.xtext.ide.editor.quickfix.DiagnosticResolutionAcceptor;
-import org.eclipse.xtext.ide.server.Document;
+import org.eclipse.xtext.ide.server.codeActions.ICodeActionService2.Options;
 
 import com.regnosys.rosetta.ide.util.CodeActionUtils;
 import com.regnosys.rosetta.rosetta.Import;
@@ -22,27 +22,30 @@ public class RosettaCodeActionProvider extends AbstractCodeActionProvider {
 	private ImportManagementService importManagementService;
 	@Inject
 	private CodeActionUtils codeActionUtils;
+	
+	@Override
+	public List<CodeAction> getCodeActions(Options options) {
+		List<CodeAction> result = new ArrayList<>();
+
+		// Handle Sorting CodeAction
+		RosettaModel model = (RosettaModel) options.getResource().getContents().get(0);
+		if (!importManagementService.isSorted(model.getImports())) {
+			result.add(codeActionUtils.createUnresolvedCodeAction("Sort imports", options.getCodeActionParams(),
+					CodeActionKind.SourceOrganizeImports));
+		}
+
+		return result;
+	}
 
 	@CodeActionResolution("Sort imports")
-	public void sortImports(DiagnosticResolutionAcceptor acceptor) {
-		acceptor.accept("Sort imports", (Diagnostic diagnostic, EObject object, Document document) -> {
-			Import importObj = (Import) object;
-			EObject container = importObj.eContainer();
+	public List<TextEdit> sortImports(RosettaModel model) {
+		EList<Import> imports = model.getImports();
 
-			if (container instanceof RosettaModel) {
-				RosettaModel model = (RosettaModel) container;
-				EList<Import> imports = model.getImports();
+		Range importsRange = codeActionUtils.getImportsRange(imports);
 
-				Range importsRange = codeActionUtils.getImportsRange(imports);
+		importManagementService.sortImports(imports);
+		String sortedImportsText = importManagementService.toString(imports);
 
-				importManagementService.sortImports(imports);
-				String sortedImportsText = importManagementService.toString(imports);
-
-				return List.of(new TextEdit(importsRange, sortedImportsText));
-			}
-
-			// if not model, return empty list of edits
-			return List.of();
-		});
+		return List.of(new TextEdit(importsRange, sortedImportsText));
 	}
 }
